Implement the actual app

+++ src/force_drawing.h
@@ -20,7 +20,7 @@ public:
   ForceDirectedDraw(Graph *const graph, int width, int height);
 
   // draws the graph on a png and saves to file "filename"
-  vector<pair<int, int>> drawGraph(const string &filename, int iterations);
+  vector<pair<float, float>> drawGraph(const string &filename, int iterations);
 
 private:
   Graph *const graph_;
@@ -29,11 +29,11 @@ private:
   // helper functions
 
   // normalize positions to be in bounds of the image
-  void normalizePositions(vector<pair<int, int>> &positions,
-                          pair<int, int> &min, pair<int, int> &max);
+  void normalizePositions(vector<pair<float, float>> &positions,
+                          pair<float, float> &min, pair<float, float> &max);
 
   // draws the positions of each vertex on a png
-  void drawPositions(const vector<pair<int, int>> &positions,
+  void drawPositions(const vector<pair<float, float>> &positions,
                      const string &filename);
 
   // calculates the repulsive force
@@ -44,10 +44,10 @@ private:
 
   // finds the difference/unit vector of 2 vectors (unit when unit == true)
   // (different when unit == false)
-  pair<float, float> unitVector(const pair<int, int> &pos1,
-                                const pair<int, int> pos2, bool unit);
+  pair<float, float> unitVector(const pair<float, float> &pos1,
+                                const pair<float, float> pos2, bool unit);
 
   // generates a random sequence of positions in the image to draw a vertices
   // (nodes) of size
-  vector<pair<int, int>> generateRandomPositions(int size);
+  vector<pair<float, float>> generateRandomPositions(int size);
 };
